voila/requirements.txt
boto3
botocore
mysql-connector-python
beautifulsoup4
